move tx_extension.h to a common place
