+++ ydb/core/tx/columnshard/data_sharing/common/ya.make
@@ -2,7 +2,6 @@ LIBRARY()
 
 PEERDIR(
     ydb/core/tx/columnshard/data_sharing/common/session
-    ydb/core/tx/columnshard/data_sharing/common/transactions
     ydb/core/tx/columnshard/data_sharing/common/context
 )
