+++ ydb/core/tx/columnshard/data_sharing/destination/transactions/tx_start_from_initiator.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <ydb/core/tx/columnshard/columnshard_impl.h>
-#include <ydb/core/tx/columnshard/data_sharing/common/transactions/tx_extension.h>
+#include <ydb/core/tx/columnshard/tablet/ext_tx_base.h>
 #include <ydb/core/tx/columnshard/data_sharing/destination/session/destination.h>
 
 namespace NKikimr::NOlap::NDataSharing {
