+++ ydb/core/tx/columnshard/data_sharing/modification/tasks/ya.make
@@ -7,7 +7,7 @@ SRCS(
 PEERDIR(
     ydb/core/tx/columnshard/data_sharing/initiator/controller
     ydb/core/tx/columnshard/data_sharing/common/session
-    ydb/core/tx/columnshard/data_sharing/common/transactions
+    ydb/core/tx/columnshard/tablet
     ydb/core/tx/columnshard/data_sharing/destination/transactions
 )
