+++ ydb/core/tx/columnshard/data_sharing/modification/transactions/tx_change_blobs_owning.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <ydb/core/tx/columnshard/columnshard_impl.h>
 #include <ydb/core/tx/columnshard/engines/writer/indexed_blob_constructor.h>
-#include <ydb/core/tx/columnshard/data_sharing/common/transactions/tx_extension.h>
+#include <ydb/core/tx/columnshard/tablet/ext_tx_base.h>
 
 namespace NKikimr::NOlap::NDataSharing {
