+++ ydb/core/tx/columnshard/data_sharing/modification/transactions/ya.make
@@ -10,6 +10,7 @@ PEERDIR(
     ydb/services/metadata/abstract
     ydb/core/tx/columnshard/data_sharing/protos
     ydb/core/tx/columnshard/blobs_action/protos
+    
     ydb/core/base
     ydb/core/tx/tiering
 )
