+++ ydb/core/tx/columnshard/data_sharing/source/transactions/tx_data_ack_to_source.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <ydb/core/tx/columnshard/columnshard_impl.h>
-#include <ydb/core/tx/columnshard/data_sharing/common/transactions/tx_extension.h>
+#include <ydb/core/tx/columnshard/tablet/ext_tx_base.h>
 #include <ydb/core/tx/columnshard/data_sharing/source/session/source.h>
 #include <ydb/core/tx/columnshard/blob_cache.h>
