+++ ydb/core/tx/columnshard/inflight_request_tracker.cpp
@@ -1,8 +1,7 @@
 #include "columnshard_impl.h"
 #include "columnshard_schema.h"
 #include "inflight_request_tracker.h"
-
-#include "data_sharing/common/transactions/tx_extension.h"
+#include "tablet/ext_tx_base.h"
 #include "engines/column_engine.h"
 #include "engines/reader/plain_reader/constructor/read_metadata.h"
 #include "hooks/abstract/abstract.h"
@@ -60,9 +59,9 @@ void TInFlightReadsTracker::AddToInFlightRequest(
 }
 
 namespace {
-class TTransactionSavePersistentSnapshots: public NOlap::NDataSharing::TExtendedTransactionBase<NColumnShard::TColumnShard> {
+class TTransactionSavePersistentSnapshots: public TExtendedTransactionBase<NColumnShard::TColumnShard> {
 private:
-    using TBase = NOlap::NDataSharing::TExtendedTransactionBase<NColumnShard::TColumnShard>;
+    using TBase = TExtendedTransactionBase<NColumnShard::TColumnShard>;
     const std::set<NOlap::TSnapshot> SaveSnapshots;
     const std::set<NOlap::TSnapshot> RemoveSnapshots;
     virtual bool DoExecute(NTabletFlatExecutor::TTransactionContext& txc, const TActorContext& /*ctx*/) override {
